+++ imdb/constants.py
@@ -3,7 +3,8 @@
 BASE_URL = "http://www.imdb.com"
 
 # SEARCH URLS
-SEARCH_BY_TITLE = "find?s=tt&ref_=fn_tt"
+SEARCH_NAME = "find?s=nm&ref_=fn_nm"
+SEARCH_TITLE = "find?s=tt&ref_=fn_tt"
 SEARCH_MOVIE = "find?s=tt&ttype=ft&ref_=fn_ft"
 SEARCH_TV = "find?s=tt&ttype=tv&ref_=fn_tv"
 SEARCH_VIDEOGAME = "find?s=tt&ttype=vg&ref_=fn_vg"